requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
